pulled conversation logic into seperate service
