+++ public/index.html
@@ -40,8 +40,8 @@
                 <div class="conversation-div" ng-repeat="conversation in currentConversations | orderBy : '-messages[messages.length-1].timestamp'">
                     {{conversation.name}}<br>
                     <span ng-repeat="userid in conversation.userids">{{getUserFromId(userid).name}} {{$last ? '' : ($index==conversation.userids.length-2) ? ' and ' : ', '}}</span>
-                    <button ng-click="setClearedForConversationMessages(conversation.id, true)">Clear conversation</button>
-                    <button ng-click="setClearedForConversationMessages(conversation.id, false)">Show all messages</button>
+                    <button ng-click="setClearedForConversationMessages(currentConversations, conversation.id, true)">Clear conversation</button>
+                    <button ng-click="setClearedForConversationMessages(currentConversations, conversation.id, false)">Show all messages</button>
                     <md-menu>
                     <md-button ng-click="$mdOpenMenu($event)">
                     Add user
