+++ public/js/ChatApp.js
@@ -4,4 +4,5 @@ angular.module("ChatApp", ["ngAnimate", "ngMaterial", "toastr"]);
 
 require("./ChatController.js");
 require("./RequestService.js");
+require("./ConversationService.js");
 require("../css/main.css");
